+++ sdppp_python/apis.py
@@ -13,6 +13,9 @@ import numpy as np
 from io import BytesIO
 from PIL import Image
 from .protocols.photoshop import ProtocolPhotoshop
+import json
+import folder_paths
+import os
 
 # str => PIL.Image
 image_cache = dict()
@@ -45,7 +48,6 @@ def registerComfyHTTPEndpoints(sdppp, PromptServer):
             image.save(stream, "PNG")
             return web.Response(body=stream.getvalue(), content_type='image/png')
         except Exception as e:
-            print('=============error============', e)
             return web.json_response({
                 'error': str(e)
             })
@@ -68,124 +70,116 @@ def registerSDHTTPEndpoints(sdppp, app):
 
 def registerSocketEvents(sdppp, sio):
 
-    # only emit by sd webui instance
-    @sio.event
-    async def c_get_image(sid, payload={}):
-        if not sdppp.has_ps_instance():
-            return
-        document = payload['document']
 
-        return await ProtocolPhotoshop.get_image(document['instance_id'], 
-            document_identify=document['identify'], 
-            layer_identify=payload['layer_identify'], 
-            boundaries=payload['boundaries'],
-        )
-
-    # only emit by sd webui instance
-    @sio.event
-    async def c_send_image(sid, payload={}):
-        if not sdppp.has_ps_instance():
-            return
-        document = payload['document']        
-
-        await ProtocolPhotoshop.send_images(document['instance_id'], 
-            document_identify=document['identify'], 
-            layer_identifies=payload['layer_identifies'], 
-            boundaries=payload['boundaries'],
-            image_urls=payload['image_urls']
-        )
 
     @sio.event
-    async def c_psd(sid, payload = {}):
-        if 'sid' not in payload:
-            return {"error": "sid is not defined"}
-        return await sdppp.sio.call('c_psd', payload, to=payload['sid'])
-
-    # only emit by photoshop instance
-    @sio.event
-    async def b_page_run(sid, payload = {}):
-        to_sid = payload['sid']
-        del payload['sid']
-        await sdppp.sio.emit('b_page_run', payload, to=to_sid)
-
-    @sio.event
-    async def b_workflow_action(sid, payload = {}):
-        if len(sdppp.page_instances) == 0:
-            return {"error": "Please connect at least one page instance"}
+    async def F_photoshop(sid, payload = {}):
+        if 'sid' not in payload or payload['sid'] not in sdppp.ppp_instances:
+            return {"error": f"sid {payload['sid']} not found"}
         
-        result = await sdppp.sio.call('b_workflow_action', payload, to=payload['sid'])
+        if payload['action'] == 'uploadImage':
+            image = payload['params']['image']
+            filename = payload['params']['filename']
+            result = await upload_image(sdppp, image, filename)
+            return result
+        else:
+            return await sdppp.sio.call('F_photoshop', payload, to=payload['sid'])
+
+    @sio.event
+    async def B_photoshop(sid, payload = {}):
+        if 'sid' not in payload or payload['sid'] not in sdppp.ppp_instances or sdppp.ppp_instances[payload['sid']].type != 'photoshop':
+            return {"error": f"sid {payload['sid']} not found"}
+        return await sdppp.sio.call('B_photoshop', payload, to=payload['sid'])
+
+    @sio.event
+    async def F_workflow(sid, payload = {}):
+        if 'sid' not in payload or payload['sid'] not in sdppp.ppp_instances:
+            return {"error": f"sid {sid} not found"}
+        result = await sdppp.sio.call('F_workflow', payload, to=payload['sid'])
         return result
 
     @sio.event
-    async def b_set_widget_value(sid, payload = {}):
-        if len(sdppp.page_instances) == 0:
-            return {"error": "Please connect at least one page instance"}
-        result = await sdppp.sio.call('b_set_widget_value', payload, to=payload['sid'])
+    async def B_workflow(sid, payload = {}):
+        if 'sid' not in payload or payload['sid'] not in sdppp.ppp_instances:
+            return {"error": f"sid {sid} not found"}
+        
+        result = await sdppp.sio.call('B_workflow', payload, to=payload['sid'])
         return result
 
     @sio.event
-    async def b_flush_data(sid, payload = {}):
-        try: 
-            store = sdppp.backend_instances[sid].store
-            if store.patch_version_acceptable(payload['fromVersion']):
-                store.patch_data(payload['operations'], payload['fromVersion'])
-            else:
-                result = await sio.call('s_request_data', {}, to=sid)
-                sdppp.backend_instances[sid].store.sync_data(result['data'], result['version'])
-        except Exception as e:
-            return {"error": str(e)}
+    async def store_flush(sid, payload = {}):
+        if sid not in sdppp.ppp_instances:
+            return {"error": f"sid {sid} not found"}
+        instance = sdppp.ppp_instances[sid]
+        store = instance.store
+
+        if store.patch_version_acceptable(payload['fromVersion']):
+            store.patch_data(payload['operations'], payload['fromVersion'])
+        else:
+            result = await sio.call('store_request', {}, to=sid)
+            store.sync_data(result['data'], result['version'])
 
         payload['sid'] = sid
-        if len(sdppp.page_instances):
-            await sio.emit('s_flush_data', payload, to=[page.sid for page in sdppp.page_instances.values()])
+        payload['type'] = instance.type
+        if len(sdppp.ppp_instances):
+            await sio.emit('store_flush', payload)
 
     @sio.event
-    async def c_flush_data(sid, payload = {}):
-        try:
-            store = sdppp.page_instances[sid].store
-            if store.patch_version_acceptable(payload['fromVersion']):
-                store.patch_data(payload['operations'], payload['fromVersion'])
-            else:
-                result = await sio.call('s_request_data', {}, to=sid)
-                sdppp.page_instances[sid].store.sync_data(result['data'], result['version'])
-        except Exception as e:
-            print('=============error============', e)
-            return {"error": str(e)}
-
-        payload['sid'] = sid
-        if len(sdppp.backend_instances):
-            await sio.emit('s_flush_data', payload, to=[backend.sid for backend in sdppp.backend_instances.values()])
-
-    @sio.event
-    async def c_request_data(sid, payload = {}):
-        if 'sid' in payload:
+    async def store_request(sid, payload = {}):
+        if 'sid' in payload: # request single
             return {
                 'sid': payload['sid'],
-                'data': sdppp.backend_instances[payload['sid']].store.data,
-                'version': sdppp.backend_instances[payload['sid']].store.version
+                'data': sdppp.ppp_instances[payload['sid']].store.data,
+                'version': sdppp.ppp_instances[payload['sid']].store.version
             }
-        else:
+
+        elif 'type' in payload: # request all 
             ret = {}
-            for backend in sdppp.backend_instances.values():
-                ret[backend.sid] = {
-                    'data': backend.store.data,
-                    'version': backend.store.version
-                }
+            for ppp in sdppp.ppp_instances.values():
+                if ppp.type == payload['type']:
+                    ret[ppp.sid] = {
+                        'data': ppp.store.data,
+                        'version': ppp.store.version
+                    }
             return ret
 
-    @sio.event
-    async def b_request_data(sid, payload = {}):
-        if 'sid' in payload:
-            return {
-                'sid': payload['sid'],
-                'data': sdppp.page_instances[payload['sid']].store.data,
-                'version': sdppp.page_instances[payload['sid']].store.version
-            }
-        else:
-            ret = {}
-            for page in sdppp.page_instances.values():
-                ret[page.sid] = {
-                    'data': page.store.data,
-                    'version': page.store.version
-                }
-            return ret
+async def upload_image(sdppp, image, filename):
+    # 找到 image_upload 处理函数
+    upload_handler = None
+    for route in sdppp.PromptServer.instance.routes:
+        if (isinstance(route, web.RouteDef) and 
+            route.method == "POST" and 
+            route.path == "/upload/image"):
+            upload_handler = route.handler
+            break
+        
+    if upload_handler is None:
+        return {"error": "Upload handler not found"}
+    
+    # 创建模拟请求对象
+    class MockFile:
+        def __init__(self, data):
+            self.file = BytesIO(data)
+            self.filename = filename
+
+    class MockPost:
+        def __init__(self, image):
+            self.data = {}
+            self.data["image"] = MockFile(image)
+            self.data["overwrite"] = "true"
+            self.data['subfolder'] = 'sdppp'
+        
+        async def post(self):
+            return self
+        
+        def get(self, key, default=None):
+            return self.data.get(key, default)
+
+    # 调用上传处理函数
+    post = MockPost(image)
+    result = await upload_handler(post)
+    
+    # 如果返回是 JSON 响应，获取实际数据
+    if isinstance(result, web.Response):
+        return json.loads(result.text)
+    pass